Crate fifo

A first-in-first-out bytes ring-buffer like kfifo in Linux.

Example

use std::io::prelude::*;
use fifo::{fifo, Sender, Receiver};

let (mut sender, mut receiver) = fifo(128);

let bytes_to_write = [0 as u8; 512];
assert_eq!(sender.write(&bytes_to_write).unwrap(), 128);

let mut bytes_to_read = [1 as u8; 512];
assert_eq!(receiver.read(&mut bytes_to_read).unwrap(), 128);

assert_eq!(bytes_to_write[0..128], bytes_to_read[0..128]);

Modules

splice
Zero-copy interface for Sender and Receiver, in splice style in Linux. When you want to write bytes into the ring-buffer from a Read, or read bytes from this into a Write, you can use SpliceRead and SpliceWrite.

Structs

Receiver
The fifo receiver. It's Send but !Send.

Sender
The fifo sender. It's Send but !Send.

Functions

align_up_for_fifo_size
Align the request size up to power of 2. If it's still less then 128 after aligned, then use 128 as return value.

fifo
Construct the fifo with capacity as align_up_for_fifo_size(usize), and return the Sender and Receiver pair connected with that.