Module fifo::splice

Zero-copy interface for Sender and Receiver, in splice style in Linux. When you want to write bytes into the ring-buffer from a Read, or read bytes from this into a Write, you can use SpliceRead and SpliceWrite.

Traits

SpliceRead
Copy some bytes directly from another Read object, without use any temporary buffers.

SpliceWrite
Copy some bytes directly into another Write object, without use any temporary buffers.